requirements.txt
Flask==3.0.0
matplotlib==3.7.0
numpy==1.23.5
pandas==2.0.1
Pillow==10.1.0
Requests==2.31.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
scikit-learn==1.0.2